+++ NEWS.md
@@ -1,3 +1,5 @@
 # heritable 0.1.0
 
 * Initial CRAN submission.
+* Broad sense heritability for single environment models (`lme4` and `asreml`)
+* See https://anu-aagi.github.io/news/posts/2026-01-12/